Create instructions module
