+++ pyqir/src/lib.rs
@@ -6,6 +6,7 @@
 mod context;
 mod evaluator;
 mod generator;
+mod instructions;
 mod python;
 mod types;
 mod utils;
