+++ pyqir/src/python.rs
@@ -4,9 +4,11 @@
 use crate::{
     evaluator::PyNonadaptiveJit,
     generator::{
-        bitcode_to_ir, ir_to_bitcode, r#const, Attribute, BasicQisBuilder, Builder, Call, FCmp,
-        FloatPredicate, ICmp, Instruction, IntPredicate, Module, Opcode, Phi, SimpleModule, Switch,
-        TypeFactory,
+        bitcode_to_ir, ir_to_bitcode, r#const, Attribute, BasicQisBuilder, Builder, Module,
+        SimpleModule, TypeFactory,
+    },
+    instructions::{
+        Call, FCmp, FloatPredicate, ICmp, Instruction, IntPredicate, Opcode, Phi, Switch,
     },
     types::{is_qubit, is_result, ArrayType, FunctionType, IntType, PointerType, StructType, Type},
     values::{
